Remove PermissionHolder from Common. Add to PermissionHolder interface
